Enabled findbugs enforcement at build-time for many packages with custom code. (#122)
Enabled findbugs enforcement at build-time for many packages with custom code. Removed support for legacy java serialization in models because it's broken for some of them. We can collect feedback for this in 2.0 so that we can decide whether we need to support this or if we can provide a better solution for the customers that need this functionality.